Refactor: seperate api request and axios
